+++ src/controllers/users.js
@@ -4,8 +4,8 @@ const bcrypt=require("bcrypt")
 
 const users=require("../models/users")
 const keys = require("../models/keys")
+const notifications=require("../models/notifications")
 const {sendMail}=require("../helpers/mail")
-const { NONAME } = require("dns")
 
 //cloudinary configuration
 cloudinary.config()
@@ -19,7 +19,7 @@ const login=async(req,res)=>{
       const validPasswd=await bcrypt.compare(password,user.password)
       if(!validPasswd) return res.status(401).json({status:"failed",message:"Wrong Email or Password"})
       const token=jwt.sign({userId:user._id},process.env.JWT_SECRET,{expiresIn:60*60})
-      res.status(200).json({status:"successfull",token})
+      res.status(200).json({status:"successfull",user:{status:user.status,role:user.role},token})
     }catch(err){
       console.log(err)
       res.status(500).json({status:"failed",message:"Internal Server Error"})
@@ -28,6 +28,9 @@ const login=async(req,res)=>{
 
 const signup=async (req, res) => {
     try {
+
+      console.log("Hello")
+
       const { name, email, password, verifyPassword } = req.body;
       if (!name || !email || !password || !verifyPassword) return res.status(400).json({status:"failed",message:"Missing Data Fields"})
       if(password !== verifyPassword) return res.status(400).json({status:"failed",message:"Passwords do not match"})
@@ -42,31 +45,35 @@ const signup=async (req, res) => {
       await sendMail(user.email,"Account Activation",`
       <html>
       <body>
-      <p>click on this <a href="${req.protocol}://${req.headers.host}/users/activate/${activationKey.key}">link</a> to activate your account</p><br>
-      <p>The activation link expires in 30 minutes after arrival.</p>
+      <p>Your code is: ${activationKey.code}</p><br>
+      <p>Use this code to activate your account</p><br>
+      <p>The code expires in 30 minutes after arrival.</p>
       </body>
       </html>
       `)
 
-      res.status(200).json({status:"successfull",message:"Signup complete, Please Activate Your Account Via The Link Sent To Your Email",activation:activationKey})
+      await notifications.create({content:`New user: ${user.name} joined the blog!`})
+
+      res.status(200).json({status:"successfull",message:"Signup complete, Please Activate Your Account Via The Code Sent To Your Email",activation:activationKey})
     } catch (err) {
       console.log(err);
       return res.status(500).json({status:"failed",message:"Internal Server Error"})
     }
 }
 
-const resendActivationLink=async(req,res)=>{
+const resendActivationCode=async(req,res)=>{
     try{
       const activationKey=await keys.create({user:req.user._id,keyType:"activation"})
       await sendMail(req.user.email,"Account Activation",`
       <html>
       <body>
-      <p>click on this <a href="${req.protocol}://${req.headers.host}/users/activate/${activationKey.key}">link</a> to activate your account</p><br>
-      <p>The activation link expires in 30 minutes after arrival.</p>
+      <p>Your code is: ${activationKey.code}</p><br>
+      <p>Use this code to activate your account</p><br>
+      <p>The code expires in 30 minutes after arrival.</p>
       </body>
       </html>
       `)
-      res.status(200).json({status:"successfull",message:"Please check your email for an activation link"})
+      res.status(200).json({status:"successfull",message:"Please check your email for an activation Code"})
     } catch (err) {
       console.log(err);
       return res.status(500).json({status:"failed",message:"Internal Server Error"})
@@ -127,16 +134,88 @@ const editUserInfo=async(req,res)=>{
   }
 }
 
-const changeUserPasswd=async(req,res)=>{
+const changeUserPassword=async(req,res)=>{
   try{
+    const {password,newPassword,verifyPassword}=req.body
+    if(!password || !newPassword || !verifyPassword) return res.status(400).json({status:"failed",message:"Missing data fields"})
+    if(newPassword !== verifyPassword) return res.status(400).json({status:"failed",message:"Passwords do not match"})
+    const user=await user.findById(req.user._id)
+    if(!user) return res.status(404).json({status:"failed",message:"No such user found"})
+    const isValidPassword=await bcrypt.compare(password,user.password)
+    if(!isValidPassword) return res.status(401).json({status:"failed",message:"Invalid Password"})
+    const salt=await bcrypt.genSalt(10)
+    const hashedPassword= await bcrypt.hash(newPassword,salt)
+    user.password=hashedPassword
+    await user.save()
+    res.status(200).json({status:200,message:"Password changed successfully"})
   }catch(err){
     console.log(err)
+    res.status(500).json({status:"failed",message:"Internal Server Error"})
+  }
+}
+
+const sendPassworResetKey=async(req,res)=>{
+  try{
+    const {email}=req.body
+    
+    if(!email) return res.status(400).json({status:"failed",message:"Missing email data field"})
+    const user=await users.findOne({email:email})
+    if(!user) return res.status(404).json({status:"failed",message:`No user with email: ${email} found`})
+    const passwdResetKey=await keys.create({user:user._id,keyType:"passwdReset"})
+    await sendMail(user.email,"Password Reset",`
+    <html>
+    <body>
+    <p>Your code is: ${passwdResetKey.code}</p><br>
+    <p>Use this code to reset your password</p><br>
+    <p>The code expires in 30 minutes after arrival.</p>
+    </body>
+    </html>
+    `)
+
+    console.log(passwdResetKey)
+    res.status(200).json({status:"successfull",message:"Successfully sent password reset key"})
+
+  }catch(err){
+    console.log(err)
+    res.status(500).json({status:"failed",message:"Internal Server Error"})
+  }
+}
+
+const resetPassword=async(req,res)=>{
+  try{
+    const {newPassword,verifyPassword}=req.body
+    const {code}=req.params
+
+    if(!newPassword || !verifyPassword) return res.status(400).json({status:"failed",message:"Missing password data fields"})
+    if(newPassword !== verifyPassword) return res.status(400).json({status:"failed",message:"Passwords do not match"})
+
+    if(!code) return res.status(400).json({status:"failed",message:"Missing password reset code"})
+    const existingkey=await keys.findOne({code:code})
+    if(!existingkey) return res.status(404).json({status:"failed",message:"Invalid Key or key does't exist"})
+
+    const user=await users.findById(existingkey.user)
+    if(!user) return res.status(404).json({status:"failed",message:"No user associated with this key was found"})
+
+    const salt=await bcrypt.genSalt(10)
+    const hashedPassword=await bcrypt.hash(newPassword,salt)
+
+    user.password=hashedPassword
+    await user.save()
+
+    await keys.findByIdAndDelete(existingkey._id)
+
+    res.status(200).json({status:200,message:"Password reset successfully"})
+
+  }catch(err){
+    console.log(err)
+    res.status(500).json({status:"failed",message:"Internal Server Error"})
   }
 }
 
 const deleteUser=async(req,res)=>{
   try{
     const {uid}=req.params
+    if(!uid) return res.status(400).json({status:"failed",message:"Missing User Id"})
     const deletedUser=await users.deleteOne(uid)
     if(!deletedUser) return res.status(404).json({status:"failed",message:"No such user"})
     res.json({status:"successful",user:deletedUser})
@@ -149,8 +228,13 @@ const deleteUser=async(req,res)=>{
 const changeUserRole=async(req,res)=>{
   try{
     const {role}=req.body
+    const {uid}=req.params
+    if(!uid) return res.status(400).json({status:"failed",message:"Missing uid parameter"})
+    if(!role) return res.status(400).json({status:"failed",message:"No role provided"})
     if(role!=="user" && role!=="admin") return res.status(400).json({status:"failed",message:"Invalid Role"})
-    const updatedUser=await users.findByIdAndUpdate(req.user._id,{role:role},{new:true})
+    const user=await users.findById(uid)
+    if(!user) return res.status(404).json({status:"failed",message:"No such user found"})
+    const updatedUser=await users.findByIdAndUpdate(uid,{role:role},{new:true})
     res.status(200).json({status:"successful",user:updatedUser})
   }catch(err){
     res.status(500).json({status:"failed",message:"Internal Server Error"})
@@ -171,12 +255,12 @@ const deactivateUser=async(req,res)=>{
 
 const activateUser=async(req,res)=>{
   try{
-    const {key}=req.params
-    if(!key) return res.status(400).json({status:"failed",message:"Missing Activation Key"})
-    const existingKey=await keys.findOne({keyType:"activation",key:key})
-    if(!existingKey) return res.status(404).json({status:"failed",message:"Invalid Or Expired Activation Key"})
+    const {code}=req.body
+    if(!code) return res.status(400).json({status:"failed",message:"Missing Activation Code"})
+    const existingKey=await keys.findOne({keyType:"activation",code:parseInt(code)})
+    if(!existingKey) return res.status(404).json({status:"failed",message:"Invalid Or Expired Activation Code"})
     const activatedUser=await users.findByIdAndUpdate(existingKey.user,{status:"active"},{new:true})
-    await keys.deleteOne({keyType:"activation",key:key})
+    await keys.deleteOne({keyType:"activation",code:parseInt(code)})
     res.status(200).json({status:"successful",user:activatedUser})
   }catch(err){
     res.status(500).json({status:"failed",message:"Internal Server Error"})
@@ -184,4 +268,4 @@ const activateUser=async(req,res)=>{
   }
 }
 
-module.exports={login,signup,getCurrentUser,getUser,getUsers,editUserInfo,deleteUser,changeUserPasswd,changeUserRole,resendActivationLink,deactivateUser,activateUser}
+module.exports={login,signup,getCurrentUser,getUser,getUsers,editUserInfo,deleteUser,changeUserRole,resendActivationCode,deactivateUser,activateUser,changeUserPassword,resetPassword,sendPassworResetKey}
